improve packaging and fix numpy deprecations
with these changes `pip install .` should work directly

+++ NPTFit/incgamma_fct.pyx
@@ -35,7 +35,7 @@ cdef np.ndarray incgamma_up_fct_ary_int(int m_max, double z, double a):
     inc_gamma_upper_ratio_ary(m_max+1, z-1, a, &result[0])
     return result
 
-def incgamma_up_fct_ary(np.int m_max, np.float z, np.float a):
+def incgamma_up_fct_ary(int m_max, float z, float a):
     """ Python interface, don't call this from Cython code """
     return incgamma_up_fct_ary_int(m_max, z, a)
 
@@ -53,6 +53,6 @@ cdef np.ndarray incgamma_lo_fct_ary_int(int m_max, double z, double a):
     inc_gamma_lower_ratio_ary(m_max+1, z-1, a, &result[0])
     return result
 
-def incgamma_lo_fct_ary(np.int m_max, np.float z, np.float a):
+def incgamma_lo_fct_ary(int m_max, float z, float a):
     """ Python interface, don't call this from Cython code """
     return incgamma_lo_fct_ary_int(m_max, z, a)
